% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{convert_se_dge}
\alias{convert_se_dge}
\title{Convenience function to convert from  \code{\link[SummarizedExperiment]{SummarizedExperiment}} to a \code{\link[edgeR]{DGEList}}}
\usage{
convert_se_dge(raw_se)
}
\arguments{
\item{raw_se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}}
}
\value{
a \code{\link[edgeR]{DGEList}}
}
\description{
Takes in a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object containing RNA-seq data and converts to \code{\link[edgeR]{DGEList}}.
}

