% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{de_test_rseq}
\alias{de_test_rseq}
\title{Perform a pair-wise differential expression test on RNA-seq data}
\usage{
de_test_rseq(dge, comparison)
}
\arguments{
\item{comparison}{A two-element array indicating the 'baseline' and 'test' classes IN THAT ORDER. DE testing  will be performed relative to baseline (element 2 vs 1).}

\item{se}{A \code{\link[edgeR]{DGEList}}}
}
\value{
the fitted and adjusted \code{\link[edgeR]{TopTags}} object
}
\description{
Takes in a \code{\link[edgeR]{DGEList}} containing (normalized) RNA-seq data, performs a fit using \code{\link[edgeR]{estimateCommonDisp}} and \code{\link[edgeR]{estimateTagwiseDisp}}, a differential expression test via \code{\link[edgeR]{exactTest}} and multiple-testing correction using Benjamini-Hochberge method in \code{\link[edgeR]{topTags}}.
}

