% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emRNASeq.R
\docType{package}
\name{emRNASeq}
\alias{emRNASeq}
\alias{emRNASeq-package}
\title{emRNASeq: A package for pathways analysis of RNA sequencing data}
\description{
emRNASeq: A package for pathways analysis of RNA sequencing data
}

