% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{filter_rseq}
\alias{filter_rseq}
\title{Filter RNA-seq data for minimumnumber of counts in a class}
\usage{
filter_rseq(se, baseline, test, min_counts = 1)
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}}

\item{baseline}{character for the 'baseline' class}

\item{test}{character for the 'test' class. DE testing will be performed relative to}

\item{min_counts}{The minimum threshold in counts per million for each sample}
}
\value{
a filtered \code{\link[edgeR]{DGEList}}
}
\description{
Takes in a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object containing RNA-seq data and filters for low counts (> 1 cpm in a minimum number of samples equal to the smallest group size)
}

