% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{format_class_gsea}
\alias{format_class_gsea}
\title{Generate categorical class text file content}
\usage{
format_class_gsea(dge, de_tested_tt)
}
\arguments{
\item{dge}{A \code{\link[edgeR]{DGEList}}, typically a filtered version coming out of  \code{\link{filter_rseq}}}

\item{tested_tt}{A \code{\link[edgeR]{TopTags}}}
}
\value{
a matrix
}
\description{
Creates content conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#CLS:_Categorical_.28e.g_tumor_vs_normal.29_class_file_format_.28.2A.cls.29}{GSEA's text file format for discrete classes}.
}

