% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{format_expression_gsea}
\alias{format_expression_gsea}
\title{Generate an expression text file content}
\usage{
format_expression_gsea(dge)
}
\arguments{
\item{dge}{A \code{\link[edgeR]{DGEList}}, typically the result of \code{\link[edgeR]{calcNormFactors}}}
}
\value{
a \code{\link[base]{data.frame}} of the expression data
}
\description{
Creates a \code{\link[base]{data.frame}} conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#TXT:_Text_file_format_for_expression_dataset_.28.2A.txt.29}{GSEA's text file format for an expression dataset}.
}

