% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{format_ranks_gsea}
\alias{format_ranks_gsea}
\title{Generate text file content for genes ranked by a funtion of p-value for differential expression}
\usage{
format_ranks_gsea(adjusted_tt)
}
\arguments{
\item{adjusted_tt}{This is the \code{\link[edgeR]{TopTags}} object emerging from \code{\link{process_rseq}}}
}
\value{
a \code{\link[base]{data.frame}} of the ranks
}
\description{
Creates content conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#RNK:_Ranked_list_file_format_.28.2A.rnk.29}{GSEA's text file format for a ranked list file (.rnk)}. The column header names are arbitrary. The gene rank is based on: \eqn{sign(log(fold_change) * -log( pvalue )}.
}

