% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc.R
\name{integrate_classes}
\alias{integrate_classes}
\title{Integrate the sample class assignments}
\usage{
integrate_classes(data_se, d_merge_key, pheno_file, p_merge_key, p_pheno_key)
}
\arguments{
\item{data_se}{\code{\link[SummarizedExperiment]{SummarizedExperiment}}}

\item{d_merge_key}{string data_se key to merge on}

\item{pheno_file}{string path to phenotypes data file}

\item{p_merge_key}{string pheno_file data column to merge on}

\item{p_pheno_key}{string pheno_file phenotype assignments}
}
\value{
The \code{\link[edgeR]{DGEList}}
}
\description{
Integrate the sample class assignments
}

