% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{make_class}
\alias{make_class}
\title{Generate categorical class text file content}
\usage{
make_class(filtered_dge, bh_adjusted_tt, filepath = ".")
}
\arguments{
\item{filtered_dge}{A \code{\link[edgeR]{DGEList}}}

\item{bh_adjusted_tt}{A \code{\link[edgeR]{TopTags}}}

\item{filepath}{a string indicating a valid local path.}
}
\description{
Creates content conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#CLS:_Categorical_.28e.g_tumor_vs_normal.29_class_file_format_.28.2A.cls.29}{GSEA's text file format for discrete classes}. Writes data relative to getwd().
}

