% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{make_class_data}
\alias{make_class_data}
\title{Generate categorical class text file content}
\usage{
make_class_data(se)
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}}
}
\value{
a list of 3 row elements; one for each line in the class file
}
\description{
Creates content conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#CLS:_Categorical_.28e.g_tumor_vs_normal.29_class_file_format_.28.2A.cls.29}{GSEA's text file format for discrete classes}
}

