% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{make_expression}
\alias{make_expression}
\title{Generate an expression text file content}
\usage{
make_expression(normalized_dge, filepath = ".")
}
\arguments{
\item{normalized_dge}{A \code{\link[edgeR]{DGEList}}}

\item{filepath}{a string indicating a valid local path.}
}
\description{
Creates a content conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#TXT:_Text_file_format_for_expression_dataset_.28.2A.txt.29}{GSEA's text file format for an expression dataset}. Writes data relative to getwd().
}

