% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{make_ranks}
\alias{make_ranks}
\title{Generate text file content for genes ranked by a funtion of p-value for differential expression}
\usage{
make_ranks(adjusted_tt, filepath = ".")
}
\arguments{
\item{adjusted_tt}{This is the \code{\link[edgeR]{TopTags}} object emerging from \code{\link{process_rseq}}}

\item{filepath}{a string indicating a valid local path.}
}
\description{
Creates content conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#RNK:_Ranked_list_file_format_.28.2A.rnk.29}{GSEA's text file format for a ranked list file (.rnk)}. The column header names are arbitrary. The gene rank is based on: \eqn{sign(log(fold_change) * -log( pvalue )}. Writes data relative to getwd().
}

