% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{make_ranks_data}
\alias{make_ranks_data}
\title{Generate text file content for genes ranked by a funtion of p-value for differential expression}
\usage{
make_ranks_data(pathname = "expression.txt", se)
}
\arguments{
\item{tmm_normalized_dge}{This is the \code{\link[edgeR]{DGEList}} of normalized RNA-seq data from \code{\link{process_rseq}}}
}
\description{
Creates content conforming to \href{http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#RNK:_Ranked_list_file_format_.28.2A.rnk.29}{GSEA's text file format for a ranked list file (.rnk)}. The column header names are arbitrary.
}

