% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{map_names}
\alias{map_names}
\title{Normalize the data.frame nameswith those on a \code{\link[GenomicRanges]{GRanges}} object}
\usage{
map_names(data_df, gene_model, source_name)
}
\arguments{
\item{data_df}{the data frame to synchronize}

\item{gene_model}{the \code{\link[GenomicRanges]{GRanges}} object to match rows with}
}
\value{
A \code{\link[base]{data.frame}} with the same rows and names as the input \code{\link[GenomicRanges]{GRanges}}
}
\description{
Assumes that the gene_model has a meta-data column with source_name
}

