% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{normalize_rseq}
\alias{normalize_rseq}
\title{Think wrapper for RNASeq normalization package of choice (edgeR)}
\usage{
normalize_rseq(filtered_dge)
}
\arguments{
\item{filtered_dge}{A \code{\link[edgeR]{DGEList}}, typically the output of \code{\link{filter_rseq}}}
}
\value{
the normalized \code{\link[edgeR]{DGEList}} object
}
\description{
Takes in a \code{\link[edgeR]{DGEList}} containing (filtered) RNA-seq data, performs a normalization using  \code{\link[edgeR]{calcNormFactors}} with default method "TMM"
}

