% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{plot_de}
\alias{plot_de}
\title{An M vs. A plot that highlights differentially expressed genes}
\usage{
plot_de(de_tested_tt, filtered_dge, baseline, test, threshold = 0.05)
}
\arguments{
\item{de_tested_tt}{A \code{\link[edgeR]{topTags}}, typically the output of \code{\link{de_test_rseq}}}

\item{filtered_dge}{A \code{\link[edgeR]{DGEList}}, typically the output of \code{\link{filter_rseq}}}

\item{baseline}{character array for the 'baseline' class}

\item{test}{character array for the 'test' class.}

\item{threshold}{the maximum value of FDR (q-value) for a gene to be considered differentially expressed}
}
\description{
Takes in an \code{\link[edgeR]{topTags}} and  \code{\link[edgeR]{DGEList}} objects and the classes that are being compared and uses the \code{\link[edgeR]{plotSmear}} to display the differentially expressed genes.
}

