% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsvnew.R
\name{readcsvnew}
\alias{readcsvnew}
\title{Read CSV file}
\usage{
readcsvnew(file, ...)
}
\arguments{
\item{file}{a csv file.}

\item{...}{arguments passed to read.csv}
}
\description{
Simple wrapper for read.csv
}

