% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc.R
\name{retrieve_gdc_rnaseq}
\alias{retrieve_gdc_rnaseq}
\title{Get TCGA RNAseq data from GDC}
\usage{
retrieve_gdc_rnaseq(project, barcodes)
}
\arguments{
\item{project}{a string of the project-code}

\item{barcodes}{a character vector of TCGA barcodes. You must
include at least the first three barcode segments (TCGA-XX-XXXX). Omit this
to download all available}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
\description{
Get TCGA RNAseq data from GDC
}

