% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{filter_rseq}
\alias{filter_rseq}
\title{Filter RNA-seq data for minimumnumber of counts in a class}
\usage{
filter_rseq(se, comparison)
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}}

\item{comparison}{A two-element array indicating the 'baseline' and 'test' classes IN THAT ORDER. DE testing  will be performed relative to baseline (element 2 vs 1).}
}
\value{
a filtered \code{\link[edgeR]{DGEList}}
}
\description{
Takes in a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object containing RNA-seq data and filters for low counts (> 1 cpm in a minimum number of samples equal to the smallest group size)
}

