% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{get_gene_model}
\alias{get_gene_model}
\title{Retrieve the gene info}
\usage{
get_gene_model(data_df, species)
}
\arguments{
\item{data_df}{the data frame of genes (rownames) and samples (colnames)}

\item{species}{the species (mouse, human)}
}
\value{
A dataframe of gene attributes
}
\description{
This requires that either the mgi_id for mouse or ensembl gene id for human are used.
This is tough to enforce so must fail elegantly
}

