% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{merge_data}
\alias{merge_data}
\title{Merge a set of HT-Seq RNA expression files in tab-delimited format with
two columns and no header. The classes are defined by a meta data file
which is a tab-delimited text file with headers for the sample read 'id' and
'class'. Each row entry is a corresponding filename and class assignment.
Only accepts pair-wise comparison so there must be exactly 2 classes.}
\usage{
merge_data(meta_file, species, ...)
}
\arguments{
\item{meta_file}{A metadata file}

\item{species}{A character array indicating the species}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
\description{
Merge a set of HT-Seq RNA expression files in tab-delimited format with
two columns and no header. The classes are defined by a meta data file
which is a tab-delimited text file with headers for the sample read 'id' and
'class'. Each row entry is a corresponding filename and class assignment.
Only accepts pair-wise comparison so there must be exactly 2 classes.
}

