% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcess.R
\name{process_rseq}
\alias{process_rseq}
\title{Perform a pair-wise differential expression analysis on the raw
RNA-seq data}
\usage{
process_rseq(raw_se, comparison)
}
\arguments{
\item{raw_se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}}

\item{comparison}{A two-element array indicating the 'baseline' and 'test' classes IN THAT ORDER. DE testing  will be performed relative to baseline (element 2 vs 1).}
}
\value{
a list of objects including the filtered (filtered_dge) and normalized \code{\link[edgeR]{DGEList}} (tmm_normalized_dge) and the adjusted \code{\link[edgeR]{TopTags}} (bh_adjusted_tt)
}
\description{
Takes in a SummarizedExperiment object containing RNA-seq data and processes
it using \code{\link{edgeR}} sequentially: 1. Filtering for low counts (> 10 cpm in some minimum number of samples); 2. Normalization via the Trimmed Mean of M-values method in \code{\link[edgeR]{calcNormFactors}}; 3. Fit using \code{\link[edgeR]{estimateCommonDisp}}  and \code{\link[edgeR]{estimateTagwiseDisp}}; 4. Differential Expression testing via \code{\link[edgeR]{exactTest}};  5. Multiple-testing correction using Benjamini-Hochberge method in \code{\link[edgeR]{topTags}}.
}

